updated tvl calculation
