+++ package.json
@@ -16,7 +16,7 @@
   "license": "ISC",
   "dependencies": {
     "@defillama/sdk": "latest",
-    "@drift-labs/sdk": "^2.33.0",
+    "@drift-labs/sdk": "^2.35.1-beta.6",
     "@project-serum/anchor": "^0.25.0",
     "@solana/web3.js": "^1.36.0",
     "@solendprotocol/solend-sdk": "^0.6.2",
