+++ projects/drift.js
@@ -1,25 +1,72 @@
-const { sumTokens2 } = require('./helper/solana')
+const { PublicKey } = require("@solana/web3.js");
+const { getTokenAccountBalances } = require("./helper/solana");
+const {
+  initialize,
+  getSpotMarketVaultPublicKey,
+  getInsuranceFundVaultPublicKey,
+} = require("@drift-labs/sdk");
+
+async function getSdkProps() {
+  const sdkConfig = initialize({ env: "mainnet-beta" });
+  const spotMarkets = sdkConfig.SPOT_MARKETS;
+
+  return {
+    spotMarkets,
+    sdkConfig,
+  };
+}
 
 async function tvl() {
-  // token account authority: JCNCMFXo5M5qwUPg2Utu1u6YWp3MbygxqBsBeXXJfrw
-  return sumTokens2({
-    tokenAccounts: [
-      '6W9yiHDCW9EpropkFV8R3rPiL8LVWUHSiys3YeW6AT6S', // legacy usdc vault
-      'Bzjkrm1bFwVXUaV9HTnwxFrPtNso7dnwPQamhqSxtuhZ', // usdc insurance fund
-      'GXWqPpjQpdz7KZw9p7f5PX2eGxHAhvpNXiviFkAB8zXg', // usdc vault
-      'DfYCNezifxAEsQbAJ1b3j6PX3JVBe8fu11KBhxsbw5d2', // sol vault
-      '4vwQWa4RjmPkn1WrmyEE3t912yWsBf9JNkASH36AQL3F', // sol insurance fund
-      '5p8B6KhJjesV212heBu1o86W2vUSnW1P83ZNnMLtCAAx', // msol vault
-      'iBM2BTsrXXDfwm4P4ssbzBAquaj7gGgNNhBVaq8ryiY',  // msol insurance fund
-      '2CqkQvYxp9Mq4PqLvAQ1eryYxebUh4Liyn5YMDtXsYci', // legacy usdc insurance fund
-    ]
-  })
+  const sdkProps = await getSdkProps();
+  const spotMarkets = sdkProps.spotMarkets;
+  const sdkConfig = sdkProps.sdkConfig;
+
+  const spotMarketVaults = await Promise.all(
+    spotMarkets.map((mkt) =>
+      getSpotMarketVaultPublicKey(
+        new PublicKey(sdkConfig.DRIFT_PROGRAM_ID),
+        mkt.marketIndex
+      )
+    )
+  );
+
+  const insuranceFundVaults = await Promise.all(
+    spotMarkets.map((mkt) =>
+      getInsuranceFundVaultPublicKey(
+        new PublicKey(sdkConfig.DRIFT_PROGRAM_ID),
+        mkt.marketIndex
+      )
+    )
+  );
+
+  const legacyVaults = [
+    new PublicKey("6W9yiHDCW9EpropkFV8R3rPiL8LVWUHSiys3YeW6AT6S"), // legacy usdc vault
+    new PublicKey("2CqkQvYxp9Mq4PqLvAQ1eryYxebUh4Liyn5YMDtXsYci"), // legacy usdc insurance fund
+  ];
+
+  const tvlBalances = await getTokenAccountBalances([
+    ...spotMarketVaults,
+    ...insuranceFundVaults,
+    ...legacyVaults,
+  ]);
+
+  const tvlOutput = Object.entries(tvlBalances).reduce(
+    (previous, [token, amount]) => {
+      return {
+        ...previous,
+        [`solana:${token}`]: amount,
+      };
+    },
+    {}
+  );
+
+  return tvlOutput;
 }
 
 module.exports = {
   timetravel: false,
   methodology: "Calculate sum across all program token accounts",
   solana: {
-    tvl
-  }
-}
+    tvl,
+  },
+};
